pyproject.toml
[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
# disabling rules conflicting with the formatter (from the docs)
select = ["E", "F", "I001", "RUF100"]
# TODO: remove F401 from ignored later
ignore = [
    "W191", "E111", "E114",
    "E117", "D206", "D300",
    "Q000", "Q001", "Q002",
    "Q003", "COM812", "COM819",
    "ISC001", "ISC002", "F401"
]

[tool.ruff.format]
skip-magic-trailing-comma = false

[project]
name = "pla-laom"
version = "0.1.0"
dependencies = [
    "torchvision",
    "torchinfo",
    "torcheval",
    "opencv-python",
    "mujoco==3.2.0",
    "numpy",
    "dm_control==1.0.3",
    "gymnasium",
    "shimmy==2.0.0",
    "imageio",
    "imageio-ffmpeg",
    "Pillow",
    "pyrallis==0.3.1",
    "h5py",
    "tqdm",
    "stable_baselines3",
    "sb3-contrib==2.4.0",
    "vector-quantize-pytorch==1.22.16",
    "wandb",
    "tensordict"
]

[tool.ruff.lint.isort]
# see https://github.com/astral-sh/ruff/issues/8571
known-third-party = ["wandb"]

[project.optional-dependencies]
dev = [
    "ruff"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


